requirements/base.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
cryptography==42.0.5
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.8
executing==2.0.1
factory-boy==3.3.0
Faker==24.0.0
fastapi==0.110.0
filelock==3.13.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
identify==2.5.35
idna==3.6
importlib_resources==6.1.2
iniconfig==2.0.0
ipython==8.22.2
jedi==0.19.1
limits==3.9.0
matplotlib-inline==0.1.6
nodeenv==1.8.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.7.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.0.2
pytest-env==1.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
shortuuid==1.0.12
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.36.3
tenacity==8.2.3
traitlets==5.14.1
typing_extensions==4.10.0
uvicorn==0.27.1
virtualenv==20.25.1
wcwidth==0.2.13
wrapt==1.16.0
